registry rename and doc updates
