+++ pallets/chain-space/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = 'pallet-registry'
-description = 'Manage Registries'
+name = 'pallet-chain-space'
+description = 'Manage Chain Spaces'
 version = "0.9.0"
 authors = ['Dhiway Networks <info@dhiway.com>']
 edition = '2021'
